server/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
